Fix eval robot spawn location, by default disable "terrain level" sampling
By default, locomotion samples the terrain for a random tile. The tile
could be anywhere, so the camera doesn't match. Instead, add an option
to spawn at tile (0,0).

We override Eval config to set this to false:

--terrain.terrain-term.randomize-spawn=False

An alternative is to snap the camera to the robot, which we can consider
in the future. This change is simpler and still useful regardless. I.e,
"terrain level" doesn't make sense for planes.

cr: https://code.amazon.com/reviews/CR-237317007