Improved robot spawning on uneven terrain when applying random XY offsets
The changes:
- Added query_terrain_heights method to query terrain height at arbitrary XY positions
- Updated robot spawning to place robots at correct height above terrain surface
- Fixed terrain generation bugs where vertical_scale wasn't properly used in two terrain functions

Previously we randomized (-1,1) for XY but didn't update Z for the new terrain height. This
will manifest for steep terrain, and isn't a (big) issue with our existing config defaults.
This still isn't perfect as we sample on robot base XY and not feet/collision bounds. Ideally
we expose these options to config but it should be a part of domain randomization, and also handle randomized
joint angles too. E.g, feet pointing down need to move up to avoid penetrating terrain on spawn.

cr: https://code.amazon.com/reviews/CR-238731346